added timelineItem component and added modal
