--- src/components/timeline/ModalTimeline.js
@@ -1,8 +1,0 @@
-import React, { useState } from 'react';
-import ReactDOM from 'react-dom';
-import 'antd/dist/antd.css';
-import { Modal } from 'antd';
-
-export default function ModalTimeline(details) {
-  return <div>{details.desc}</div>;
-}
