+++ src/components/timeline/RecentTimeline.js
@@ -1,64 +1,18 @@
 import React, { useState } from 'react';
-import { Timeline, TimelineItem } from 'vertical-timeline-component-for-react';
-import { DateTime } from 'luxon';
-import { nanoid } from 'nanoid';
+import { Timeline } from 'vertical-timeline-component-for-react';
 import { useTimeline } from '../../state/query_hooks/useTimeline';
-import ModalTimeLine from './ModalTimeline';
 import './RecentTimeline.css';
+import TimelineItems from './TimelineItems';
 
 export function RecentTimeline() {
   const timelineQuery = useTimeline();
-
-  // parses url to get website name
-  const urlDomain = url => {
-    let re = /^(?:https?:\/\/)?(?:[^@\n]+@)?(?:www\.)?([^:/\n?]+)/gim;
-    let newUrl = url.split(re)[1].replace('.com', '');
-    return newUrl;
-  };
 
   return timelineQuery.isSuccess ? (
     <div className="timeline-container">
       <h1> Timeline of Recent Events </h1>
       <Timeline lineColor={'#</Timeline>ddd'}>
         {timelineQuery.data.map(details => {
-          return (
-            <TimelineItem
-              key={nanoid()}
-              description={details.desc}
-              dateText={DateTime.fromISO(details.date).toLocaleString(
-                DateTime.DATE_FULL
-              )}
-              style={{ color: '#BC541E' }}
-              dateInnerStyle={{ color: 'white', backgroundColor: '#003767' }}
-            >
-              <h3>{details.title}</h3>
-              <h4 className="cityState">
-                {details.city}, {details.state}
-              </h4>
-              <div className="categories">
-                {details.categories.map(element => (
-                  <span className="category-item" key={nanoid()}>
-                    {element}
-                  </span>
-                ))}
-              </div>
-
-              <div className="timeline-links">
-                <h4 className="sourceText"> Sources: </h4>
-                {details.src.map(element => (
-                  <a
-                    href={element}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    className="link-button"
-                    key={nanoid()}
-                  >
-                    {urlDomain(element)}
-                  </a>
-                ))}
-              </div>
-            </TimelineItem>
-          );
+          return <TimelineItems details={details} />;
         })}
       </Timeline>
     </div>
